requirements.txt
annotated-types
anyio
beautifulsoup4
certifi
charset-normalizer
click
fastapi
greenlet
h11
idna
mariadb
packaging
pydantic
pydantic_core
PyMySQL
python-dotenv
requests
sniffio
soupsieve
SQLAlchemy
starlette
typing_extensions
urllib3
uvicorn
